backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
oracledb==1.4.2
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
